Implements numpy array earlier in CreateGuesses
CreateGuesses still parses the parameters from the input file into 2D lists of strings, but those lists are added to a numpy array of determined size and recast as floats.  I'm hoping this could eliminate the need for all of the list transformation functions currently used.  The tricky functions to reconfigure are going to be compile_initial_guess() and createRandomGuess().